Some improvements to building Soup via MSVC

+++ CLI/cli_3d.cpp
@@ -140,7 +140,7 @@ static void render()
 	console.setForegroundColour(255, 0, 255);
 	std::string info = "[WASDRF] Position  [↑←↓→] Rotation  [QE] FOV  [H] Save hi res image";
 	console << info;
-	console.setCursorPos(0, 1 + (unicode::utf8_char_len(info) / c.width));
+	console.setCursorPos(0, static_cast<unsigned int>(1 + (unicode::utf8_char_len(info) / c.width)));
 	console << "frame time: " << (render_time + print_time) << "ms (" << render_time << "ms render, " << print_time << "ms print)";
 }
